:recycle: Make account card of account home screen a separated component
